MD5 checker
Introducing files checker. At the moment, this is an optional application to use. In the future, the use of the MD5 checker will be embedded in the application itself and it will not be possible to launch it without checking hashes. Now you can use this application to check the integrity of application files